feat: Refactor useLogin hook to include driver information
The commit refactors the useLogin hook in the useLogin.ts file to include driver information. This change allows the hook to return the driver object along with the token in the LoginResponse. The updated hook is used in the AuthStore to authenticate users and store the driver information in the DriverStore.